Refactor TotalTopList
