+++ src/types.ts
@@ -1,0 +1,8 @@
+export interface Vote {
+  key?: string;
+  user: string;
+  artist: string;
+  music: number;
+  performance: number;
+  clothes: number;
+}
