+++ src/utils.ts
@@ -1,5 +1,5 @@
-import { Vote } from "./Voting";
+import { Vote } from "./types";
 
 export const calculateVoteScore = (vote: Vote) => {
-    return vote.music * 3 + vote.performance * 2 + vote.clothes * 1;
-}
+  return vote.music * 3 + vote.performance * 2 + vote.clothes * 1;
+};
